Fix typo in command to log pip-diff failure
The `pip-diff` tool from vendor/pip-pop is used to determine stale
requirements. When `pip-diff` encounters an unexpected failure, a count
is logged using `mcount` from heroku/buildpack-stdlib, and compilation
is continued.

Due to a typo, mount(8) was invoked instead of `mcount`, with an invalid
argument. Under `set +e`, this error would abort compilation instead of
continuing it.