Do not run all tests when no tests match TESTS
Define a noop test function when TESTS is passed and no function in the
test script matches the contents of TESTS. The noop function merely
prints a line notifying the user that the tests were skipped.

This happens because tests were split up into multiple test scripts, but
the Makefile passes TESTS to all of them. When the `suite` hook does not
invoke `suite_addTest`, shunit2 falls back to running all defined tests.